refactor: move modify commits to packed_scalar module
